third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/html/semantics/scripting-1/the-script-element/module/instantiation-error-5-expected.txt

This is a testharness.js-based test.
FAIL Test that loading a graph in which a module is already errored results in that module's error. assert_array_equals: property 2, expected object "SyntaxError: The requested module does not provide an export named 'something'" but got object "SyntaxError: The requested module does not provide an export named 'something'"
Harness: the test ran to completion.
